ProtocolPolicy.Factory -> ProtocolPolicyFactory
